feat: 🎸 add delete news and contents
