+++ src/app/admin/@dumper/page.tsx
@@ -2,8 +2,8 @@ import { TableSkeleton } from "@/components/table/table-skeleton";
 import { Separator } from "@/components/ui/separator";
 import { Unauthorized } from "@/components/unauthorized";
 import { checkAdminPermission } from "@/features/auth/utils/role";
-import { ContentsTable } from "@/features/update-status/components/contents-table";
-import { NewsTable } from "@/features/update-status/components/news-table";
+import { ContentsTableProvider } from "@/features/update-status/components/contents-table-provider";
+import { NewsTableProvider } from "@/features/update-status/components/news-table-provider";
 import { Suspense } from "react";
 
 export const dynamic = "force-dynamic";
@@ -17,12 +17,12 @@ export default async function Page() {
 				<div className="space-y-2">
 					<h2 className="px-4 pt-4">NEWS</h2>
 					<Suspense fallback={<TableSkeleton />}>
-						<NewsTable />
+						<NewsTableProvider />
 					</Suspense>
 					<Separator className="h-px bg-gradient-to-r from-primary to-primary-grad" />
 					<h2 className="px-4 pt-4">CONTENTS</h2>
 					<Suspense fallback={<TableSkeleton />}>
-						<ContentsTable />
+						<ContentsTableProvider />
 					</Suspense>
 				</div>
 			) : (
