+++ src/features/update-status/components/news-table.tsx
@@ -1,55 +1,124 @@
-"use server";
+"use client";
 import { StatusCodeView } from "@/components/status-code-view";
-import { ERROR_MESSAGES } from "@/constants";
-import { hasAdminPermissionOrThrow } from "@/features/auth/utils/get-session";
+import { TableFooter } from "@/components/table/table-footer";
+import {
+	Dialog,
+	DialogContent,
+	DialogDescription,
+	DialogHeader,
+	DialogTitle,
+} from "@/components/ui/dialog";
+import {
+	Table,
+	TableBody,
+	TableCell,
+	TableHead,
+	TableHeader,
+	TableRow,
+} from "@/components/ui/table";
+import { DeleteNewsButton } from "@/features/dump/components/delete-news-button";
 import type { NewsAtom } from "@/features/dump/stores/news-atom";
-import { DumpTable } from "@/features/update-status/components/dump-table";
-import { loggerError } from "@/pino";
-import prisma from "@/prisma";
+import { newsColumns } from "@/features/update-status/components/news-columns";
+import {
+	type ColumnFiltersState,
+	type SortingState,
+	type VisibilityState,
+	flexRender,
+	getCoreRowModel,
+	getFilteredRowModel,
+	getPaginationRowModel,
+	getSortedRowModel,
+	useReactTable,
+} from "@tanstack/react-table";
+import { useState } from "react";
 
-export async function NewsTable() {
-	try {
-		await hasAdminPermissionOrThrow();
+type Props = {
+	data: NewsAtom[];
+};
 
-		const news = await prisma.news.findMany({
-			select: {
-				id: true,
-				title: true,
-				quote: true,
-				url: true,
-				status: true,
-				Category: { select: { name: true } },
-			},
-		});
+export function NewsTable({ data }: Props) {
+	const [sorting, setSorting] = useState<SortingState>([]);
+	const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
+	const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({});
 
-		return (
-			<DumpTable<NewsAtom>
-				data={news.map((d) => {
-					return {
-						id: d.id,
-						title: d.title,
-						quote: d.quote,
-						url: d.url,
-						status: d.status,
-						category: d.Category.name,
-					};
-				})}
-				columnType="NEWS"
+	const [open, setOpen] = useState(false);
+	const [dialogData, setDialogData] = useState<NewsAtom>();
+
+	const table = useReactTable({
+		data,
+		columns: newsColumns(),
+		onSortingChange: setSorting,
+		onColumnFiltersChange: setColumnFilters,
+		getCoreRowModel: getCoreRowModel(),
+		getPaginationRowModel: getPaginationRowModel(),
+		getSortedRowModel: getSortedRowModel(),
+		getFilteredRowModel: getFilteredRowModel(),
+		onColumnVisibilityChange: setColumnVisibility,
+		state: { sorting, columnFilters, columnVisibility },
+	});
+
+	const handleClick = (rowData: NewsAtom) => {
+		setDialogData(rowData);
+		setOpen(true);
+	};
+
+	return (
+		<div className="p-4">
+			<Table className="min-w-[1080px] overflow-x-scroll">
+				<TableHeader>
+					{table.getHeaderGroups().map((headerGroup) => (
+						<TableRow key={headerGroup.id}>
+							{headerGroup.headers.map((header) => {
+								return (
+									<TableHead key={header.id}>
+										{header.isPlaceholder
+											? null
+											: flexRender(
+													header.column.columnDef.header,
+													header.getContext(),
+												)}
+									</TableHead>
+								);
+							})}
+						</TableRow>
+					))}
+				</TableHeader>
+				<TableBody>
+					{table.getRowModel().rows?.length ? (
+						table.getRowModel().rows.map((row) => (
+							<TableRow key={row.id} onClick={() => handleClick(row.original)}>
+								{row.getVisibleCells().map((cell) => (
+									<TableCell key={cell.id}>
+										{flexRender(cell.column.columnDef.cell, cell.getContext())}
+									</TableCell>
+								))}
+							</TableRow>
+						))
+					) : (
+						<TableRow>
+							<TableCell colSpan={newsColumns.length} className="mx-auto flex">
+								<StatusCodeView statusCode="204" />
+							</TableCell>
+						</TableRow>
+					)}
+				</TableBody>
+			</Table>
+			<TableFooter
+				numberOfRows={table.getFilteredRowModel().rows.length}
+				onClickPrevious={() => table.previousPage()}
+				previousButtonDisabled={!table.getCanPreviousPage()}
+				onClickNext={() => table.nextPage()}
+				nextButtonDisabled={!table.getCanNextPage()}
 			/>
-		);
-	} catch (error) {
-		loggerError(
-			ERROR_MESSAGES.UNEXPECTED,
-			{
-				caller: "NewsTable",
-				status: 500,
-			},
-			error,
-		);
-		return (
-			<div className="flex flex-col items-center">
-				<StatusCodeView statusCode="500" />
-			</div>
-		);
-	}
+			<Dialog open={open} onOpenChange={setOpen}>
+				<DialogContent>
+					<DialogHeader>
+						<DialogTitle>{dialogData?.title ?? ""}</DialogTitle>
+						<DialogDescription>{dialogData?.quote ?? ""}</DialogDescription>
+					</DialogHeader>
+					<DeleteNewsButton id={dialogData?.id} />
+				</DialogContent>
+			</Dialog>
+		</div>
+	);
 }
